+++ packages/react/src/components/FormField/src/components/FormFieldInput/index.ts
@@ -1,4 +1,5 @@
 import { FormFieldInput } from "./FormFieldInput";
 
 export { FormFieldInput } from "./FormFieldInput";
+export type { FormFieldInputProps } from "./FormFieldInput";
 export default FormFieldInput;