+++ packages/react/src/components/FormField/src/components/FormFieldSelect/index.ts
@@ -1,4 +1,5 @@
 import { FormFieldSelect } from "./FormFieldSelect";
 
 export { FormFieldSelect } from "./FormFieldSelect";
+export type { FormFieldSelectProps } from "./FormFieldSelect";
 export default FormFieldSelect;