+++ packages/react/src/components/FormField/src/components/FormFieldTextarea/index.ts
@@ -1,4 +1,5 @@
 import { FormFieldTextarea } from "./FormFieldTextarea";
 
 export { FormFieldTextarea } from "./FormFieldTextarea";
+export type { FormFieldTextareaProps } from "./FormFieldTextarea";
 export default FormFieldTextarea;