+++ packages/react/src/components/FormField/src/index.ts
@@ -2,4 +2,7 @@ import { FormField } from "./FormField";
 
 export { FormField } from "./FormField";
 export type { FormFieldProps } from "./formField.types";
+export type { FormFieldInputProps } from "./components/FormFieldInput";
+export type { FormFieldSelectProps } from "./components/FormFieldSelect";
+export type { FormFieldTextareaProps } from "./components/FormFieldTextarea";
 export default FormField;